Add first pass ECMH implementation in blockchain package
The plan as of now is to use ECMH (Elliptic Curve Multiset Hash) to keep a
running hash commitment of the state of the validator set. ECMH is used here
instead of a merkle tree because we'd need to recalculate the entire merkle
tree every time we want to remove a validator. ECMH allows for efficient
removals though it is slow overall. For small validators sets simply recalculating
a merkle tree would be faster. But once the validator set gets over around 64 or
so then ECMH should be faster.

We dont *need* to commit the hash of validator set to the block headers, it's
just a nice to have feature that would allow people to fast sync the state
at a given block header.

This implementation is based on old work done by Thomas Van der Wassen for BCH.
As such it needs to be evaluated to see if we can do it better. It's still a
work in progress and needs thorough testing.