Create blockindex and ValidatorSet implementations
The blockindex is used to track the state of the chain in the database and
create a mechanism for traversing the chain backwards and forwards. This implementation
is mostly complete and has a test.

The ValidatorSet implementation is still a bit of a work in progress and needs
testing. The goal here is to store the full validator set in memory and periodically
flush changes to disk.