Create pruned blockchain feature
When a chain is pruned the node will delete all but the 10 most recent blocks
from disk. The buffer of 10 is kept so as to minimize the chance that the wallet
will get out of sync if the node crashes before the wallet commits a block.

This commit also extends the useragent field to advertise services that the node
supports. At this point the only service will be the blockchain service.

Full nodes offer the blockchain service, pruned nodes do not.

The sync manager package will only try to sync from peers advertising the
blockchain service.