Refactor validator reward and stake slashing
This change makes it so the block production limit is evaluated every block
rather than just once per epoch.

If any given block pushes a validator over the limit they lose their block
reward and earn a strike. Three strikes and they lose their stake.

The limit is not reset after each strike so if a block pushes a validator over
the limit and they proceed to keep creating blocks they will lose their stake
in just two more blocks.

This change requried refactoring the validator set code because this change means
that we need to update the nullifier set *after* updating the validator set
because we need to know which, if any, nullifiers from banned validators must
be added to the nullifier set.

This is extra problematic because of how everything is supposed to be committed
atomically. This commit creates a new validator set transaction object allow us
to compute the banned nullifiers before the final commit.