Check number of validator connections before finalizing blocks
This commit adds a new ValidatorConnector class that manages maintaining
connections to the validators.

The consensus engine is also modified to check if we are connected to >50% of
the validator set before sending out queries.

The purpose of this is to prevent a node (that has stake) from finalizing blocks
by itself and forking itself off the network in the event it can't make outgoing
connections.